go1.26

+++ utils/alioss/oss.go
@@ -362,15 +362,13 @@ func (c *ossClient) UploadFileBytes(key string, bs []byte, isPrivate bool, force
 		logger.Error("[alioss] error:" + err.Error())
 		return fileName, err
 	}
-	if ex {
-		if forceUpload {
-			err = c.C.DeleteObject(key)
-			if err != nil {
-				logger.Error("[alioss] error:" + err.Error())
-			}
-		} else {
-			return key, err
+	if forceUpload && ex {
+		err = c.C.DeleteObject(key)
+		if err != nil {
+			logger.Error("[alioss] error:" + err.Error())
 		}
+	} else {
+		return key, err
 	}
 	err = c.C.PutObject(key, bytes.NewReader(bs), contentType, acl)
 	if err != nil {
